src/requirements.txt
fastapi
uvicorn[standard]
websockets
python-multipart
pillow
matplotlib==3.6.2
imageio==2.23.0
numpy==1.24.1
gdown==4.6.0
opencv-python==4.6.0.66
tensorflow==2.10.1